Add header to all pages and clean up code
